practicoSQL

+++ Angular/src/app/components/Entidades/proveedores/proveedores.component.ts
@@ -9,21 +9,29 @@ import { ProveedoresService } from 'src/app/service/proveedores.service';
 })
 export class ProveedoresComponent {
   proveedores: Proveedor[] = [];
-  
+
   constructor(public serv: ProveedoresService) {}
 
   ngOnInit(): void {
     this.serv.getDatos().subscribe((data: Proveedor[]) => {
       this.proveedores = data;
     });
-
   }
 
   public deleteProveedor(id: number) {
-    this.serv.deleteProveedor(id);
-    this.serv.getDatos().subscribe((data: Proveedor[]) => {
-      this.proveedores = data;
-    });
+    const confirmDelete = window.confirm(
+      '¿Estás seguro de que quieres eliminar este proveedor?'
+    );
+
+    if (confirmDelete) {
+      // Si el usuario confirma, proceder con la eliminación
+      this.serv.deleteProveedor(id);
+
+      // Actualizar la lista de proveedores después de la eliminación
+      this.serv.getDatos().subscribe((data: Proveedor[]) => {
+        this.proveedores = data;
+      });
+    }
   }
   public editProveedor(id: number): void {
     this.serv.getProveedorById(id).subscribe((proveedor) => {
